refactor: refactored image upload component
