+++ web/src/components/rhf/file-upload/index.tsx
@@ -1,13 +1,9 @@
-import { Box, Button, IconButton, Stack, TextField } from "@mui/material"
-import { CloudinaryImage } from "@cloudinary/url-gen"
-import { FunctionComponent, useState } from "react"
+import { Box, Button, Stack, TextField } from "@mui/material"
+import { FunctionComponent } from "react"
 import { Controller, useFormContext } from "react-hook-form"
 import RenderList from "@components/render-list"
-import Render from "@components/render"
-import RenderIcon from "@components/render-icon"
-import icons from "@/icons"
-import { AdvancedImage } from "@cloudinary/react"
 import useImage from "@hooks/use-image"
+import ImageMiniPreview from "./components/mini-image-preview"
 
 type Props = {
     label: string
@@ -76,79 +72,6 @@ const RHFImageUpload: FunctionComponent<Props> = ({ name, label }) => {
                 }}
             />
         </>
-    )
-}
-
-const ImageMiniPreview = ({ image }: { image: CloudinaryImage | string }) => {
-    const [onHover, setOnHover] = useState(false)
-    return (
-        <Box
-            sx={{
-                height: 100,
-                width: 100,
-                position: "relative"
-            }}
-            onMouseEnter={() => {
-                setOnHover(true)
-            }}
-            onMouseLeave={() => {
-                setOnHover(false)
-            }}
-        >
-            <Render
-                when={image !== ""}
-                show={
-                    <>
-                        <Render
-                            when={onHover}
-                            show={
-                                <Box
-                                    sx={{
-                                        position: "absolute",
-                                        top: 0,
-                                        bottom: 0,
-                                        left: 0,
-                                        right: 0,
-                                        borderRadius: "15px",
-                                        backgroundColor: "#c1c1c13b",
-                                        cursor: "pointer",
-                                        display: "flex",
-                                        justifyContent: "center",
-                                        alignItems: "center"
-                                    }}
-                                >
-                                    <IconButton size='small'>
-                                        <RenderIcon icon={icons.delete} />
-                                    </IconButton>
-                                </Box>
-                            }
-                        />
-                        <AdvancedImage cldImg={image as CloudinaryImage} />
-                    </>
-                }
-                otherwise={
-                    <Box
-                        sx={{
-                            position: "absolute",
-                            top: 0,
-                            bottom: 0,
-                            left: 0,
-                            right: 0,
-                            borderRadius: "15px",
-                            backgroundColor: "#c1c1c111",
-                            cursor: "pointer",
-                            display: "flex",
-                            justifyContent: "center",
-                            alignItems: "center"
-                        }}
-                    >
-                        <IconButton size='small'>
-                            <RenderIcon icon={icons.animated.loading} />
-                        </IconButton>
-                    </Box>
-                }
-            />
-        </Box>
     )
 }
